add tests to getScenarios
